Shop is vulnerable to challenges

+++ api.php
@@ -1,20 +1,80 @@
 <?php
 include 'functions.php';
+ini_set('session.cookie_samesite', 'Strict');
+ini_set('session.cookie_httponly', 1);
+
+if (isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on') {
+    ini_set('session.cookie_secure', 1);
+}
 session_start();
+
+#SQLMAP Trigger
+if (str_contains($_SERVER['HTTP_USER_AGENT'], "sqlmap")) {
+    print("You dont really need SqlMap for this :)");
+    http_response_code(403); // Return a forbidden status code
+    exit;
+}
+
+//---------------------------------------------------- Vulnerable Code----------------------------------------------------
+
+//simulated legacy api
+if (isset($_POST["createUser"])) {
+    if (empty($_POST["username"]) && empty($_POST["password"]) && empty($_POST["password2"]) && empty($_POST["email"])) {
+        print("Fill out all fields.");
+        exit();
+    }
+
+    $unsafe_username = $_POST["username"];
+    $unsafe_email = $_POST["email"];
+    //Wird später gehasht
+    $password = $_POST["password"];
+    $password2 = $_POST["password2"];
+
+    $username = sanitize_input($unsafe_username);
+    $email = sanitize_input($unsafe_email);
+    $role = sanitize_input($_POST["role"]);
+
+    if ($unsafe_email != $email) {
+        $error = "Email contains unallowed characters.";
+    }
+
+    if ($unsafe_username != $username) {
+        $error = "Username contains unallowed characters.";
+    }
+
+    if ($password != $password2) {
+        $error = "Passwords do not match.";
+    }
+
+    if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+        $error = "Incorrect E-Mail format.";
+    }
+
+    if ($role !== "user" && $role !== "admin") {
+        $error = "Unknown role.";
+    }
+
+    if (empty($error)) {
+        create_user($username, $password, $email, $role);
+        print("created user");
+        exit();
+    } else {
+        print($error);
+        exit();
+    }
+}
 
 //Start admin area
 if ($_SESSION["role"] === 'admin') {
-
     // Hinzufügen von Produkten
     if (isset($_POST["addProduct"])) {
         $name = sanitize_input($_POST["name"]);
         $desc = sanitize_input($_POST["description"]);
         $quantity = sanitize_input($_POST["quantity"]);
-        $image_path = basename(sanitize_input($_POST["image"]));
-
+        $image_path = sanitize_input($_POST["image"]);
         if (!empty($name) && !empty($desc) && !empty($quantity) && !empty($image_path)) {
             if (is_numeric($quantity) && $quantity >= 0) {
-                addProduct($name, $desc, $quantity, "/webshop/uploads/" . $image_path);
+                addProduct($name, $desc, $quantity, "uploads/" . $image_path);
             } else {
                 $_SESSION["adderror"] = "Invalid Quantity.";
             }
@@ -23,21 +83,29 @@ if ($_SESSION["role"] === 'admin') {
         }
         header("Location:admin_products.php");
         exit();
-    } // Löschen von Produkten
-    else if (isset($_POST["delete"])) {
-        deleteProduct(sanitize_input($_POST["delete"]));
+    }
+}
+
+#-----------------------------------------------------------------------Vulnerable Code End -------------------
+
+//Start admin area
+if ($_SESSION["role"] === 'admin') {
+    if (isset($_POST["adminProducts"])) {
+        // Löschen von Produkten
+        if (isset($_POST["delete"])) {
+            deleteProduct(sanitize_input($_POST["delete"]));
+        } //Restock
+        else if (isset($_POST["restock"])) {
+            restock();
+        } //File Upload
+        else if (isset($_FILES["fileToUpload"])) {
+            uploadFile($_FILES["fileToUpload"]);
+        } // Löschen von Uploads
         header("Location:admin_products.php");
         exit();
-    } else if (isset($_POST["restock"])) {
-        restock();
-        header("Location:admin_products.php");
-        exit();
-    } else if (isset($_FILES["fileToUpload"])) {
-        uploadFile($_FILES["fileToUpload"]);
-        header("Location:admin_products.php");
-        exit();
-    } // Löschen von Uploads
-    else if (isset($_POST["delete_image"])) {
+    }
+
+    if (isset($_POST["delete_image"])) {
         deleteImage(sanitize_input($_POST["delete_image"]));
         header("Location:admin_products.php");
         exit();
@@ -72,13 +140,11 @@ if ($_SESSION["role"] === 'admin') {
 
         if (empty($error)) {
             create_user($username, $password, $email, $role);
-            header("Location:admin_users.php");
-            exit();
         } else {
             $_SESSION["error"] = $error;
-            header("Location:admin_users.php");
-            exit();
         }
+        header("Location:admin_users.php");
+        exit();
     } // Löschen von Benutzer
     else if (isset($_POST["delete_id"])) {
         deleteUser(sanitize_input($_POST["delete_id"]));
@@ -158,6 +224,7 @@ if (!isset($_SESSION["role"])) {
 
         $user = loginUser($name, $pass);
         if ($user !== "None") {
+            session_regenerate_id();
             $_SESSION["username"] = $user["username"];
             $_SESSION["role"] = $user["role"];
             $_SESSION["user_id"] = $user["id"];
